# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShareDirStat API
  version: "1"
  summary: Disk usage analysis for mounted shares, with delete and download.
  description: |
    The JSON API behind the ShareDirStat web UI. Every endpoint is also
    intended for scripts.

    **No authentication.** Access control is the job of the surrounding
    infrastructure (reverse proxy, VPN, ingress middleware).

    **Mutating requests** (`POST`) must carry the header
    `X-Requested-With: ShareDirStat` and a JSON body. Cross-origin browser
    requests are refused: no CORS headers are ever emitted, and a
    `Sec-Fetch-Site` other than `same-origin`/`none` is rejected.

    **Paths** are share-relative, `/`-separated, without a leading slash; the
    empty string is the share root. File names on Linux are bytes, not text:
    a name that is not valid UTF-8 is returned with replacement characters in
    `name` and its raw bytes base64-encoded in `name_b64`.

    **Sizes** are bytes. Each node carries both `size` (apparent, `st_size`)
    and `alloc` (allocated, `st_blocks * 512`); the `basis` query parameter
    selects which one governs sorting and percentages.

    **Caching.** Tree, treemap, top, extension and search responses carry an
    `ETag` derived from the share's current generation and mutation counter;
    `If-None-Match` returns `304`.
  license:
    name: MIT
    url: https://github.com/StevenGann/ShareDirStat/blob/main/LICENSE
servers:
  - url: /api/v1
...
paths:
...
  /shares/{id}/treemap:
    get:
      tags: [browse]
      summary: Pruned subtree for treemap rendering
      description: |
        The subtree at `path`, pruned so that no more than `max_nodes` nodes
        are emitted, none deeper than `max_depth`, and none smaller than
        `min_fraction` of the root. Every directory whose children were
        pruned carries a `truncated` summary of what was left out, so the
        drawn areas always add up to the parent.
      operationId: treemap
      parameters:
        - { $ref: "#/components/parameters/shareId" }
        - { $ref: "#/components/parameters/path" }
        - { $ref: "#/components/parameters/basis" }
        - name: min_fraction
          in: query
          schema: { type: number, minimum: 0, maximum: 1, default: 0.0005 }
        - name: max_nodes
          in: query
          schema: { type: integer, minimum: 1, maximum: 50000, default: 10000 }
        - name: max_depth
          in: query
          schema: { type: integer, minimum: 1, maximum: 64, default: 8 }
      responses:
        "200":
          description: The pruned tree.
          content:
            application/json:
              schema:
                type: object
                properties:
                  share: { type: string }
                  generation: { type: string }
                  basis: { $ref: "#/components/schemas/Basis" }
                  root: { $ref: "#/components/schemas/TreemapNode" }
                  nodes: { type: integer, description: Nodes emitted. }
                  min_fraction: { type: number }
                  max_depth: { type: integer }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/NotScanned" }
...
  /shares/{id}/extensions:
    get:
      tags: [browse]
      summary: Space by file extension
      operationId: extensions
      parameters:
        - { $ref: "#/components/parameters/shareId" }
        - { $ref: "#/components/parameters/path" }
      responses:
        "200":
          description: Extensions, largest first. `ext` is empty for files with none.
          content:
            application/json:
              schema:
                type: object
                properties:
                  extensions:
                    type: array
                    items:
                      type: object
                      properties:
                        ext: { type: string }
                        files: { type: integer }
                        size: { type: integer }
                        alloc: { type: integer }
        "409": { $ref: "#/components/responses/NotScanned" }
...
components:
  parameters:
    shareId:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9][a-z0-9_-]{0,63}$" }
...
    path:
      name: path
      in: query
      description: Share-relative path; empty for the root.
      schema: { type: string, default: "" }
    basis:
      name: basis
      in: query
      schema: { $ref: "#/components/schemas/Basis" }
  responses:
...
    NotFound:
      description: No such share, scan or path.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotScanned:
      description: The share has no results yet (`share_not_scanned`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Basis:
      type: string
      enum: [apparent, allocated]
      default: apparent
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: A stable machine-readable code.
              examples: [share_not_found, path_not_found, csrf_rejected, confirmation_required, delete_disabled]
            message: { type: string }
            details: { type: object }
...
    Node:
      type: object
      properties:
        name: { type: string }
        name_b64:
          type: string
          description: Raw bytes of a name that is not valid UTF-8.
        path: { type: string }
        kind: { type: string, enum: [dir, file, symlink, other] }
        size: { type: integer, description: Apparent size, or aggregate for a directory. }
        alloc: { type: integer, description: Allocated size, or aggregate for a directory. }
        mtime: { type: string, format: date-time }
        mode: { type: string, description: Octal, e.g. "0644". }
        perms: { type: string, description: ls-style, e.g. "-rw-r--r--". }
        uid: { type: integer }
        gid: { type: integer }
        owner: { type: string }
        group: { type: string }
        ext: { type: string }
        flags:
          type: object
          properties:
            partial: { type: boolean, description: Some entries beneath could not be read. }
            mountpoint: { type: boolean, description: Another filesystem; not scanned. }
            hardlink_dup: { type: boolean, description: Another name for an inode already counted. }
            excluded: { type: boolean, description: Entries beneath were skipped by an exclusion pattern. }
            unscanned: { type: boolean }
        files: { type: integer, description: Aggregate, directories only. }
        dirs: { type: integer, description: Aggregate, directories only. }
        children: { type: integer, description: Direct children, directories only. }
        pct_of_parent: { type: number }
        pct_of_share: { type: number }
    TreemapNode:
      allOf:
        - { $ref: "#/components/schemas/Node" }
        - type: object
          properties:
            children_list:
              type: array
              items: { $ref: "#/components/schemas/TreemapNode" }
            children_total: { type: integer }
            truncated:
              type: object
              description: Children not emitted, so their area can still be drawn.
              properties:
                children: { type: integer }
                size: { type: integer }
                alloc: { type: integer }
...